+++ front/src/components/pages/About/AboutValueList.tsx
